+++ model/datiModelli.h
@@ -13,11 +13,11 @@ class datiModelli
 private:
     QString nomeModello;
     QString tessuto;
-    float tessUsato;
-    float costoTessutoMq;
-    float costoBase;
-    float costoLavaggio;
-    float costoVendita;
+    double tessUsato;
+    double costoTessutoMq;
+    double costoBase;
+    double costoLavaggio;
+    double costoVendita;
     unsigned int produzioneGiornaliera;
 public:
     /**
@@ -31,27 +31,27 @@ public:
      * @param cv perzzo di vendita
      * @param pg produzione giornaliera del modello
      */
-    explicit datiModelli(QString n,QString t, float tu, float ctm, float cb, float cl,float cv, unsigned int pg);
+    explicit datiModelli(QString n,QString t, double tu, double ctm, double cb, double cl,double cv, unsigned int pg);
 
     //getter
     const QString& getNomeModello() const;
     const QString& getTessuto() const;
-    float getTessUsato() const;
-    float getCostoTessutoMq() const;
-    float getCostoBase() const;
-    float getCostoLavaggio() const;
-    float getCostoFinale() const;
-    float getCostoVendita() const;
+    double getTessUsato() const;
+    double getCostoTessutoMq() const;
+    double getCostoBase() const;
+    double getCostoLavaggio() const;
+    double getCostoFinale() const;
+    double getCostoVendita() const;
     unsigned int getProduzioneGiornaliera() const;
 
     //setter
     void setNomeModello(const QString& n) {nomeModello=n;}
     void setTessuto(const QString& t) {tessuto=t;}
-    void setTessutoUsato(const float tu) {tessUsato=tu;}
-    void setCostoTessutoMq(const float ctm) {costoTessutoMq=ctm;}
-    void setCostoBase(const float cb) {costoBase=cb;}
-    void setCostoLavaggio(const float cn) {costoLavaggio=cn;}
-    void setCostoVendita(const float cv){costoVendita=cv;}
+    void setTessutoUsato(const double tu) {tessUsato=tu;}
+    void setCostoTessutoMq(const double ctm) {costoTessutoMq=ctm;}
+    void setCostoBase(const double cb) {costoBase=cb;}
+    void setCostoLavaggio(const double cn) {costoLavaggio=cn;}
+    void setCostoVendita(const double cv){costoVendita=cv;}
     void setProduzioneGiornaliera(const unsigned int pg){produzioneGiornaliera=pg;}
 
 };
